Improvements to core subpackage
