+++ youtubesearchpython/core/__init__.py
@@ -1,1 +1,2 @@
 from .video import VideoCore
+from .constants import *
